R object file resource client

R object file resource client

R object file resource client

Format

A R6 object of class RDSFileResourceClient

Details

Connects to a RDS file and loads the serialized object. Similar to the R data file resource, except that the RDS format stores a single R object.

Super classes

resourcer::ResourceClient -> resourcer::FileResourceClient -> RDSFileResourceClient

Methods

Public methods

RDSFileResourceClient$new()

RDSFileResourceClient$asDataFrame()

RDSFileResourceClient$getValue()

RDSFileResourceClient$clone()

Inherited methods

resourcer::ResourceClient$asTbl()

resourcer::ResourceClient$exec()

resourcer::ResourceClient$getConnection()

resourcer::ResourceClient$getResource()

resourcer::FileResourceClient$close()

resourcer::FileResourceClient$downloadFile()

Method new()

Creates a new RDSFileResourceClient instance.

Usage

RDSFileResourceClient$new(resource)

Arguments

resource

A valid resource object.

Returns

A RDSFileResourceClient object.

Method asDataFrame()

Coerce the resource value extracted from the R object file to a data.frame.

Usage

RDSFileResourceClient$asDataFrame(...)

Arguments

...

Additional parameters to as.data.frame (not used yet).

Returns

A data.frame.

Method getValue()

Get the resource value extracted from the R object file.

Usage

RDSFileResourceClient$getValue(...)

Arguments

...

Additional parameters to get the value object (not used yet).

Returns

The resource value.

Method clone()

The objects of this class are cloneable with this method.

Usage

RDSFileResourceClient$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.