R data file Resource resolver

R data file Resource resolver

R data file Resource resolver

Format

A R6 object of class RDataFileResourceResolver

Details

The resource is a R data file and data format is the class of the symbol that will be loaded.

Super class

resourcer::ResourceResolver -> RDataFileResourceResolver

Methods

Public methods

RDataFileResourceResolver$isFor()

RDataFileResourceResolver$newClient()

RDataFileResourceResolver$clone()

Inherited methods

resourcer::ResourceResolver$initialize()

Method isFor()

Check that the provided resource has a URL that locates a R data file: the resource can be accessed as a file and the resource URL path ends with ".rda" or ".rdata" (case ignored), or the resource format is prefixed with "r:" or "rda:" (a kind of namespace to qualify the R object class).

Usage

RDataFileResourceResolver$isFor(x)

Arguments

x

The resource object to validate.

Returns

A logical.

Method newClient()

Creates a RDataFileResourceClient instance from provided resource.

Usage

RDataFileResourceResolver$newClient(x)

Arguments

x

A valid resource object.

Returns

A RDataFileResourceClient object.

Method clone()

The objects of this class are cloneable with this method.

Usage

RDataFileResourceResolver$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.