Apache Spark DBI resource connector

Apache Spark DBI resource connector

Apache Spark DBI resource connector

Format

A R6 object of class SparkResourceConnector

Details

Makes a Apache Spark connection object, that is also a DBI connection object, from a resource description.

Super class

resourcer::DBIResourceConnector -> SparkResourceConnector

Methods

Public methods

SparkResourceConnector$new()

SparkResourceConnector$isFor()

SparkResourceConnector$createDBIConnection()

SparkResourceConnector$closeDBIConnection()

SparkResourceConnector$clone()

Inherited methods

resourcer::DBIResourceConnector$getTableName()

resourcer::DBIResourceConnector$readDBTable()

resourcer::DBIResourceConnector$readDBTibble()

Method new()

Create a SparkResourceConnector instance.

Usage

SparkResourceConnector$new()

Returns

A SparkResourceConnector object.

Method isFor()

Check if the provided resource applies to a Apache Spark server. The resource URL scheme must be one of "spark", "spark+http" or "spark+https".

Usage

SparkResourceConnector$isFor(resource)

Arguments

resource

The resource object to validate.

Returns

A logical.

Method createDBIConnection()

Creates a DBI connection object from a Apache Spark resource.

Usage

SparkResourceConnector$createDBIConnection(resource)

Arguments

resource

A valid resource object.

Returns

A DBI connection object.

Method closeDBIConnection()

Close the DBI connection to Apache Spark.

Usage

SparkResourceConnector$closeDBIConnection(conn)

Arguments

conn

A DBI connection object.

Method clone()

The objects of this class are cloneable with this method.

Usage

SparkResourceConnector$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.