HTTP file resource getter

HTTP file resource getter

HTTP file resource getter

Format

A R6 object of class HttpFileResourceGetter

Details

Access a file that is stored at a HTTP(S) address. Use Basic authentication header if both resource's identity and secret are defined.

Super class

resourcer::FileResourceGetter -> HttpFileResourceGetter

Methods

Public methods

HttpFileResourceGetter$new()

HttpFileResourceGetter$isFor()

HttpFileResourceGetter$downloadFile()

HttpFileResourceGetter$clone()

Inherited methods

resourcer::FileResourceGetter$extractFileName()

resourcer::FileResourceGetter$makeDownloadDir()

Method new()

Creates a new HttpFileResourceGetter instance.

Usage

HttpFileResourceGetter$new()

Returns

A HttpFileResourceGetter object.

Method isFor()

Check that the provided resource has a URL that locates a file accessible through "http" or "https".

Usage

HttpFileResourceGetter$isFor(resource)

Arguments

resource

The resource object to validate.

Returns

A logical.

Method downloadFile()

Download the file from the remote address in a temporary location. Applies Basic authentication if credentials are provided in the resource.

Usage

HttpFileResourceGetter$downloadFile(resource, ...)

Arguments

resource

A valid resource object.

...

Unused additional parameters.

Returns

The "resource.file" object.

Method clone()

The objects of this class are cloneable with this method.

Usage

HttpFileResourceGetter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.