R data file resource client

R data file resource client

R data file resource client

Format

A R6 object of class RDSFileResourceClient

Details

Connects to a R data file and loads it in a controlled environment.

Super classes

resourcer::ResourceClient -> resourcer::FileResourceClient -> RDataFileResourceClient

Methods

Public methods

RDataFileResourceClient$new()

RDataFileResourceClient$asDataFrame()

RDataFileResourceClient$getValue()

RDataFileResourceClient$clone()

Inherited methods

resourcer::ResourceClient$asTbl()

resourcer::ResourceClient$exec()

resourcer::ResourceClient$getConnection()

resourcer::ResourceClient$getResource()

resourcer::FileResourceClient$close()

resourcer::FileResourceClient$downloadFile()

Method new()

Creates a new RDataFileResourceClient instance.

Usage

RDataFileResourceClient$new(resource)

Arguments

resource

A valid resource object.

Returns

A RDataFileResourceClient object.

Method asDataFrame()

Coerce the resource value extracted from the R data file to a data.frame.

Usage

RDataFileResourceClient$asDataFrame(...)

Arguments

...

Additional parameters to as.data.frame (not used yet).

Returns

A data.frame.

Method getValue()

Get the resource value extracted from the R data file.

Usage

RDataFileResourceClient$getValue(...)

Arguments

...

Additional parameters to get the value object (not used yet).

Returns

The resource value.

Method clone()

The objects of this class are cloneable with this method.

Usage

RDataFileResourceClient$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.